Remembered to update all the edited files.  Should now be in sync... will have to test it though.
git-svn-id: https://pure-data.svn.sourceforge.net/svnroot/pure-data/trunk@3092 df7eb9d6-776f-4174-a42d-0ffd8d5d5b5d

+++ pd/src/d_soundfile.c
@@ -30,7 +30,7 @@ objects use Posix-like threads.  */
 /***************** soundfile header structures ************************/
 
 typedef unsigned short uint16;
-typedef unsigned long uint32;
+typedef unsigned int uint32;
 
 #define FORMAT_WAVE 0
 #define FORMAT_AIFF 1
